Focused on scenario 2
Made it more like scenario two now where the titles get outputted to the
text file, also has the option to save to text file depending on input
to the prompt for which file to save in.

Needs error checking for input box and need to make it "smart" and
"predictable". Can do so by storing last response for the input box and
having it default to that or something. May need an alert box to pop up.